Update README.md
🔄 **Update [July 2025]:**
An earlier version of this project noted that "Healthy Food" was the highest-rated category. Upon further review, this was based on only one review and does not represent a reliable average. A revised analysis now excludes categories with fewer than 5 entries to ensure statistical significance. Thanks to this refinement, Desert ,Pulao and Pakistani  food categories now appear as top performers based on meaningful sample sizes.